+++ src/main/java/pl/videoaggregate/repositories/VideoLinkRepository.java
@@ -3,5 +3,5 @@ package pl.videoaggregate.repositories;
 import org.springframework.data.repository.CrudRepository;
 import pl.videoaggregate.domain.VideoLink;
 
-public interface VideoLinkRepository extends CrudRepository<VideoLink, Integer>{
+public interface VideoLinkRepository extends CrudRepository<VideoLink, Long>{
 }
