+++ src/main/java/pl/videoaggregate/services/VideoLinkService.java
@@ -6,7 +6,7 @@ import pl.videoaggregate.domain.VideoLink;
 public interface VideoLinkService {
     Iterable<VideoLink> listAllProducts();
 
-    VideoLink getProductById(Integer id);
+    VideoLink getProductById(Long id);
 
     VideoLink saveProduct(VideoLink product);
 }
